refactor: Redis TTL을 이용한 실시간 혼잡도 정합성 개선
